+++ hogame.php
@@ -1,8 +1,12 @@
 <?php
 include 'functions.php';
+$lightlist = array(21,3,19);
+
+foreach($lightlist as $id){
+	game_on($id);
+}
 
 oneOff(18);
-game_on(21);
 
 if(strpos($_SERVER['HTTP_USER_AGENT'], 'WebKit')) header('Location: ' . $_SERVER['hogame.php']);
 ?>