+++ index.php
@@ -81,6 +81,7 @@
     <ul id="ho" title="Lights">
         <li class="group">Home Office</li>
         <li><a href="hoon.php">On Bright</a></li>
+        <li><a href="howltog.php">Work light toggle</a></li>
 	    <li><a href="horlx.php">On Relax</a></li>
         <li><a href="hogame.php">Game Mode</a></li>
         <li><a href="hosp2.php">SP2 Mode</a></li>